final merges commit

+++ server-side/controllers/user.controller.js
@@ -1,7 +1,8 @@
 import { populate } from 'dotenv';
-import Users from '../models/user.model.js';
+import mongoose  from 'mongoose';
 import bcrypt from 'bcrypt';
-export const getUsers = async (req, res) => {
+import Users from '../models/user.model.js'
+export const getUsers = async (req, res,next) => {
   try {
     const users = await Users.find().populate('visitsWebsites.websiteId  profiles.blockedSites profiles.limitedWebsites.websiteId' )
     .select('-__v')
@@ -12,13 +13,15 @@ export const getUsers = async (req, res) => {
   }
 };
 
-export const getUserById = async (req, res) => {
+export const getUserById = async (req, res,next) => {
+  const id = req.params.id;
+
+  if(!mongoose.Types.ObjectId.isValid(id))
+    return next({message:'id is not valid'})
   try {
-    const id = req.params.id;
     const user = await Users.findById(id).populate('visitsWebsites profiles preferences').select('-__v');
     if (!user) {
         return next({message:'user not found ',status:404})
-      return;
     }
     res.send(user);
   } catch (err) {
@@ -28,15 +31,12 @@ export const getUserById = async (req, res) => {
 };
 
 
-export const addUser = async (req, res) => {
+export const addUser = async (req, res,next) => {
   try {
-    const { name, email, password } = req.body;
-    if (!name || !email || !password) {
-      return res.status(400).send('Missing required fields');
-    }
+    
     if (req.file ) 
-     req.body.profileImage  =req.file.originalname;
-    req.body.password= await bcrypt.hash(password, 10);
+     req.body.profileImage=req.file.originalname;
+     req.body.password= await bcrypt.hash(password, 10);
     const newUser = new Users(req.body);
     console.log('newUser',newUser);
     await newUser.save();
@@ -49,9 +49,11 @@ export const addUser = async (req, res) => {
 
 
 
-export const deleteUser = async (req, res) => {
+export const deleteUser = async (req, res,next) => {
+  const id = req.params.id;
+  if(!mongoose.Types.ObjectId.isValid(id))
+    return next({message:'id is not valid'})
   try {
-    const id = req.params.id;
     const user = await Users.findByIdAndDelete(id);
     if (!user) {
       return next({message:'user not found ',status:404})
@@ -60,18 +62,19 @@ export const deleteUser = async (req, res) => {
     res.send('User deleted successfully!');
   } catch (err) {
     console.error(err);
-   return res.status(500).send('Error deleting user');
+    next({message:err.message})
   }
 };
 
-export const updatedUser = async (req, res) => {
+export const updatedUser = async (req, res,next) => {
+  const id = req.params.id;
+  if(!mongoose.Types.ObjectId.isValid(id))
+    return next({message:'id is not valid'})
   try {
-    const id = req.params.id;
-    // const { name, email, password } = req.body;
-    // const updateFields = { name, email, password };
-    if (req.file) {
+    
+    if (req.file) 
       req.body.profileImage = req.file.originalname;
-    }
+  
     const updatedUser = await Users.findByIdAndUpdate(id, req.body, { new: true });
     if (!updatedUser) {
       return next({message:'user not found ',status:404})
@@ -79,7 +82,7 @@ export const updatedUser = async (req, res) => {
     res.status(200).json(updatedUser);
   } catch (err) {
     console.error(err);
-    res.status(500).send('Error updating user');
+    next({message:err.message})
   }
 };
